Rename Extensions.cs to ConfigurationExtensions.cs and move AutoOrganizeOptionsFactory into it's own file
